python-app/requirements-dev.txt
# Dependencias de desarrollo y testing — no incluir en producción
# Instalar con: pip install -r requirements-dev.txt
# Luego ejecutar una vez: playwright install

# Testing base
pytest==8.3.5
pytest-django==4.9.0

# E2E browser testing
pytest-playwright==0.7.0
playwright==1.50.0
